fix test

+++ src/integration/java/nz/cri/gns/NZSHM22/util/NZSHM22_InversionDiagnosticsReportBuilder_IntegrationTest.java
@@ -38,9 +38,11 @@ public class NZSHM22_InversionDiagnosticsReportBuilder_IntegrationTest {
 		
 		for (File folder : tmp_folders) {
 			File[] files = folder.listFiles();
-			for (File f : files) {
-				f.delete();
-			}		
+			if (files != null) {
+			    for (File f : files) {
+				    f.delete();
+			    }
+			}
 			Files.deleteIfExists(folder.toPath());
 		}
 	}
